+++ documentation/docs/release-notes.md
@@ -7,6 +7,7 @@ title: Release notes&#58;
 
 - `InitializableObject`: the number of attempts and the last attempt time are tracked and checked with the max attempts and the min time interval between attempts (for the CAS implementation)
 - `InitializableObject`: allow re-initialization
+- Refactor the `CasOAuthWrapperClient` configuration
 
 **v5.2.1**: